pyproject.toml
[project]
name = "video-manager"
version = "0.1.0"
description = ""
authors = [
    {name = "soojiinleee",email = "soojiin.leee@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (==4.0.1)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "redis (>=6.0.0,<7.0.0)",
    "aioredlock (>=0.7.3,<0.8.0)",
    "setuptools (>=80.4.0,<81.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "celery (>=5.5.2,<6.0.0)",
    "asgiref (>=3.8.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"

